requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contrib==0.3.0
cryptography==3.3.1
cycler==0.10.0
datasketch==1.5.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-model-utils==4.1.1
django-mptt==0.11.0
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-simpleui==2021.1.1
django-taggit==1.3.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
idna==2.10
install==1.3.4
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
joblib==0.16.0
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.6.1
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient @ file:///C:/Users/ZWH/Downloads/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.5
notifications==0.3.2
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzbar==0.1.8
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlitepy==0.1.2
sqlparse==0.3.1
swapper==1.1.2.post1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
tushare==1.2.62
urllib3==1.25.11
virtualenv==20.0.35
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
